+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Remove support for annotating generated artifacts with `javax.annotation.Generated` prior to Java 9 as we no longer support Java prior to Java 9.
+
 ### [v0.56](https://github.com/realityforge/proton/tree/v0.56) (2022-04-27) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.55...v0.56)
 
 Changes in this release: